+++ src/hyp3_volcsarvatory_monitoring/etc/entrypoint.sh
@@ -1,4 +1,0 @@
-#!/bin/bash --login
-set -e
-conda activate hyp3-volcsarvatory-monitoring
-exec python -um hyp3_volcsarvatory_monitoring "$@"